remove the username/password options, not used
`piaware-config` [literally doesn't have an option for username or password anymore](https://discussions.flightaware.com/t/password-problem/18918). That can be seen in the code but also in the startup log:

    warning: cannot set option 'flightaware-user', it is not a known config option
    warning: cannot set option 'flightaware-password', it is not a known config option
    Set allow-mlat to yes in /etc/piaware.conf:7
    Set feeder-id to xxx in /etc/piaware.conf:8

I've removed all references to that in the script and the docs. I've also updated the docs, especially around current terminology for getting the Feeder ID.